services/aircraft-service/requirements.txt
# services/aircraft-service/requirements.txt
# Aircraft Service Dependencies

# Django Framework
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-filter>=23.0
django-cors-headers>=4.0
djangorestframework-simplejwt>=5.3.0
drf-nested-routers>=0.93.4

# Database
psycopg2-binary>=2.9.9

# Caching
django-redis>=5.4.0
redis>=5.0.0

# Task Queue (optional)
celery>=5.3.0

# Message Brokers (optional, based on configuration)
# pika>=1.3.0  # RabbitMQ
# kafka-python>=2.0.0  # Kafka

# Monitoring & Logging
structlog>=23.0.0
python-json-logger>=2.0.0

# HTTP Client
httpx>=0.25.0
requests>=2.31.0

# Utilities
python-dateutil>=2.8.0
python-dotenv>=1.0.0

# Validation
pydantic>=2.0.0

# Testing
pytest>=7.4.0
pytest-django>=4.5.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
factory-boy>=3.3.0
faker>=19.0.0

# Development
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0
mypy>=1.5.0
django-stubs>=4.2.0
djangorestframework-stubs>=3.14.0

# Production
gunicorn>=21.0.0
uvicorn>=0.23.0
